Fix and big update
Roles are now properly implemented

 Every new user will be registered as a User and only be able to see the data of the data table.

  Admin user created, can execute every CRUD command in the website's datatable.

Login and register pages have been fixed and are now working properly. Bug was due to files not being in right place

Fixed session erros by adding it to the services.

Every functionality that was supposed to be added to the website, is now added.